Added better error handling
On the completion of an order Attendize Generates a ticket and sends it via email. However if the ticket could not be generated we still attempt to send a file that does not exist which causes the checkout process to hang and throw an error message. Chained the Generate Ticket and SendOrderTickets events so that if a ticket cannot be generate we don't attempt to send it.